initial/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.7
django-appconf==1.0.4
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.0.1
django-extensions==2.2.9
django-imagekit==4.0.2
idna==2.9
myapp==0.1.dev0
pilkit==2.0
Pillow==7.1.2
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
